% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bland_altman.R
\name{plot_bland_altman}
\alias{plot_bland_altman}
\title{Plot the Bland-Altman plot}
\usage{
plot_bland_altman(
  data,
  rater1,
  rater2,
  trial_var,
  id_var,
  use_colour = TRUE,
  show_unadjusted = FALSE,
  scale1_name = "Scale 1",
  scale2_name = "Scale 2",
  plot_title = "Bland–Altman Plot"
)
}
\arguments{
\item{data}{Dataset}

\item{rater1}{Scores from the first rater}

\item{rater2}{Scores from the first rater}

\item{trial_var}{Column indiciating the trial each participant is from}

\item{id_var}{Participant ID}
}
\description{
Plot the Bland-Altman plot
}
\details{
Plot the Bland-Altman plot
}
\examples{
# basic Bland–Altman plot using the included example data
plot_bland_altman(
  data      = example,
  rater1    = Scale1,
  rater2    = Scale2,
  trial_var = Trial,
  id_var    = Participant_id
)

}
